feat: implement hero statistics retrieval and update hero data structure
