+++ package.json
@@ -34,7 +34,7 @@
     "lazysizes": "^5.3.2",
     "lucide-react": "^0.216.0",
     "mongoose": "^7.2.3",
-    "next": "^14.2.3",
+    "next": "^14.2.17",
     "next-auth": "^4.24.7",
     "node-cron": "^3.0.2",
     "nodemailer": "^6.9.9",
